book-import-tool/requirements.txt
requests==2.31.0
aiohttp==3.9.1
fake-useragent==1.4.0
openpyxl==3.1.2
pandas==2.1.4
tqdm==4.66.1
python-dotenv==1.0.0
pytest==8.3.5
